refactor(phaser): organize code by feature. Move floorplan components to src/app/floorplan/scenes/, shared Phaser infrastructure to phaser/, and search-specific code to src/app/search/.
