+++ src/app/floorplan/scenes/Room.ts
